: docker容器直接使用最新发布的backtest包，不再从本地安装
此外，还修改了配置文件传递方式，强制要求映射本地配置文件。

+++ docker/rootfs/root/zillionare/entrypoint.sh
@@ -1,4 +1,11 @@
 #!/bin/sh
 
+if [ ! -f /config/defaults.yaml ]; then
+    echo "config file not found, exiting"
+    return
+fi
+
+echo "port passed through envar: $PORT" > /var/log/backtest/backtest.log
+
 cp /config/defaults.yaml ~/zillionare/backtest/config/defaults.yaml
 python3 -m backtest.app start $PORT
